backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for chat-real-estate"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
django = "^4.2.2"
langchain = "^0.0.208"
django-environ = "^0.10.0"
django-rest-framework = "^0.1.0"
llama-index = "^0.6.30"
google-auth = "^2.20.0"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.9.0"
django-cors-headers = "^4.1.0"
djangorestframework = "^3.14.0"
tiktoken = "^0.4.0"
djangorestframework-simplejwt = "^5.2.2"
pypdf = "^3.12.0"
psycopg2 = "^2.9.6"
dj-database-url = "^2.0.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
gunicorn = "^20.1.0"
docx2txt = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
